Replace old implementation
I've renamed the parquet_import.py to replace the previous implementation in byc_taxi_dag.py
